inScribble/requirements.txt
azure-functions
requests
pymongo